Extract other utils functions to other places
