+++ src/lib/utils.ts
@@ -9,96 +9,4 @@ export function toArray<T>(toConvert: T | T[]): T[] {
   return toReturn.filter((value: T) => !!value);
 }
 
-/**
- * Convenience function for uppercase a string that could be nullish.
- *
- * @param value {string}   The string value; could be null
- * @returns     {string}   The string value uppercased
- */
-export function toUpperCase(value: string): string {
-  return (value || "").toUpperCase();
-}
-
-/**
- * Convenience function for making an enum value, that could be nullish, user friendly i.e. underscores removed and sentence-cased.
- *
- * @param value {string}   The string value; could be null
- * @returns     {string}   The string value made user friendly
- */
-export function makeEnumValueUserFriendly(value: string): string {
-  if (value) {
-    value = value.replace(/_/g, " ");
-    return sentenceCase(value);
-  }
-  return value;
-}
-
-/**
- * Given the provided string is a number, pads the value with zeros until it reaches the target length.
- *
- * @param value        {string}   The number as a string
- * @param targetLength {number}   The length of the resulting string once the provided value has been padded
- * @returns            {string}   The string value with leading zeros prefixed
- */
-export function padNumberWithLeadingZeros(
-  value: string,
-  targetLength = 2
-): string {
-  const valueAsNumber = parseInt(value);
-  if (isNaN(valueAsNumber)) {
-    return value;
-  }
-
-  return value.padStart(targetLength, "0");
-}
-
-export function referenceNumber(chars: string, length: number): string {
-  let mask = "";
-  if (chars.indexOf("a") > -1) mask += "abcdefghijklmnopqrstuvwxyz";
-  if (chars.indexOf("A") > -1) mask += "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-  if (chars.indexOf("#") > -1) mask += "0123456789";
-  if (chars.indexOf("!") > -1) mask += "~`!@#$%^&*()_+-={}[]:\";'<>?,./|\\";
-  let result = "";
-  for (let i = length; i > 0; --i)
-    result += mask[Math.floor(Math.random() * mask.length)];
-  return result;
-}
-
-export function joinStrings(strings: Array<string>): string {
-  const output = [];
-  for (let i = 0; i < strings.length; i++) {
-    if (strings[i]) output.push(strings[i]);
-  }
-  return output.join(", ");
-}
-
-export function sentenceCase(string: string): string {
-  return string === undefined
-    ? ""
-    : string[0].toUpperCase() + string.slice(1).toLowerCase();
-}
-
-export function formatUrlQueryParams(
-  url: string,
-  queryParamMap: Record<string, any>
-): string {
-  const formatUrl = (queryParam, value) => {
-    if (!url.includes(queryParam)) {
-      const queryStringCombiner = url.includes("?") ? "&" : "?";
-      url = `${url}${queryStringCombiner}${queryParam}=${value}`;
-    }
-  };
-
-  Object.keys(queryParamMap).forEach((queryParam) => {
-    const value = queryParamMap[queryParam];
-    formatUrl(queryParam, value);
-  });
-
-  return url;
-}
-
-export function stringToBoolean(value: string): boolean {
-  return value ? value.toLocaleLowerCase() === "true" : false;
-}
-
 export type Callback<T> = (t: T) => void;
