+++ src/pages/register-a-beacon/beacon-information.tsx
@@ -17,7 +17,7 @@ import { FormManager } from "../../lib/form/formManager";
 import { Validators } from "../../lib/form/validators";
 import { FormSubmission } from "../../lib/formCache";
 import { FormPageProps, handlePageRequest } from "../../lib/handlePageRequest";
-import { padNumberWithLeadingZeros } from "../../lib/utils";
+import { padNumberWithLeadingZeros } from "../../lib/writingStyle";
 
 interface DateInputProps {
   monthValue: string;
