+++ src/pages/register-a-beacon/check-your-answers.tsx
@@ -20,8 +20,11 @@ import {
   IRegistration,
 } from "../../lib/registration/types";
 import { PageURLs } from "../../lib/urls";
-import { makeEnumValueUserFriendly, sentenceCase } from "../../lib/utils";
-import { useRankString } from "../../lib/writingStyle";
+import {
+  makeEnumValueUserFriendly,
+  sentenceCase,
+  useRankString,
+} from "../../lib/writingStyle";
 
 interface CheckYourAnswersProps {
   registration: IRegistration;
