+++ src/useCases/submitRegistration.ts
@@ -1,5 +1,4 @@
 import { IAppContainer } from "../lib/appContainer";
-import { referenceNumber } from "../lib/utils";
 
 export type SubmitRegistrationFn = (
   submissionId: string,
@@ -43,3 +42,15 @@ export const submitRegistration =
       referenceNumber: registration.getRegistration().referenceNumber || "",
     };
   };
+
+const referenceNumber = (chars: string, length: number): string => {
+  let mask = "";
+  if (chars.indexOf("a") > -1) mask += "abcdefghijklmnopqrstuvwxyz";
+  if (chars.indexOf("A") > -1) mask += "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+  if (chars.indexOf("#") > -1) mask += "0123456789";
+  if (chars.indexOf("!") > -1) mask += "~`!@#$%^&*()_+-={}[]:\";'<>?,./|\\";
+  let result = "";
+  for (let i = length; i > 0; --i)
+    result += mask[Math.floor(Math.random() * mask.length)];
+  return result;
+};
